+++ app/src/androidTest/java/dev/davidfdev/stela/ui/NoteListBannerTest.kt
@@ -30,6 +30,7 @@ class NoteListBannerTest {
                     onTogglePin = {},
                     onToggleSelection = {},
                     onClearSelection = {},
+                    onToggleSelectAll = {},
                     onBatchTogglePin = {},
                     onBatchDelete = {},
                     onSearchChange = {},